@@ -3,8 +3,12 @@
 #
 # Examples:
 #
-#   cities = City.create([{ :name => 'Chicago' }, { :name => 'Copenhagen' }])
-#   Mayor.create(:name => 'Daley', :city => cities.first)
+#   cities = City.create([{ name: 'Chicago' }, { name: 'Copenhagen' }])
+#   Mayor.create(name: 'Emanuel', city: cities.first)
+
+def seed_dir
+    ENV['SEED_DIR'] || File.dirname(__FILE__) + '/seeds'
+end
 
 # Create example users
 puts "Creating admins ..."
@@ -14,7 +18,6 @@ Admin.create do |a|
   a.password_confirmation = 'test'
   a.email = 'admin@example.com'
 end
-
 
 puts "Creating users ..."
 for i in 1..10 do
@@ -29,18 +32,38 @@ for i in 1..10 do
   puts "Created user #{i}"
 end
 
-# Create example maps
-puts "Creating maps ..."
-
-map_ids = File.open("#{::Rails.root.to_s}/db/map-ids.txt")
-map_ids.each_with_index do |id, index|
-  id.chomp!
-  break if index == 30
-  Map.create do |map|
-    map.title = "Map #{index+1}"
-    map.description = "This is a sample description"
-    map.tileset_url = "http://samos.mminf.univie.ac.at/maps/#{id}"
-    map.user = User.first
+# Create sample maps, either seed from YAML file in maphub-portal/db/seeds/seeddata.yaml
+seed_file = "#{seed_dir}/maps.yaml"
+if File.exists?(seed_file)
+  YAML.load_documents(File.open(seed_file, "r")) do |entry|
+    Map.create do |map|
+      map.identifier  = entry["id"]
+      map.title       = entry["title"]
+      map.subject     = entry["subject"]
+      map.author      = entry["author"]
+      map.date        = entry["date"]
+    end
+    puts "Created map #{entry['map']}"
   end
-  puts "Created map #{index+1}"
+# or from dummy variables
+else
+  maps = ["g3200.ct001129", 
+          "g3201b.ct002662",
+          "g3300.ar000600",
+          "g3300.ar003300",
+          "g3300.ct000912",
+          "ct002033"]
+
+  puts "Creating sample maps"
+
+  maps.each_with_index do |id, index|
+    Map.create do |map|
+      map.identifier  = id
+      map.title       = "Map #{index+1}"
+      map.subject     = "Lorem ipsum dolor sit amet. Sunt in culpa qui officia deserunt mollit anim id est laborum."
+      map.author      = "Anonymous"
+      map.date        = "1877"
+    end
+    puts "Created map #{id}"
+  end
 end
