@@ -10,13 +10,12 @@ class AnnotationsController < ApplicationController
   # GET /map/:map_id/annotations
   # GET /map/:map_id/annotations.xml  
   def index
-    
     unless @parent.nil?
       @annotations = @parent.annotations
     else
       @annotations = Annotation.all
     end
-      
+    
     respond_to do |format|
       format.html # index.html.erb
       format.xml  { render :xml => @annotations }
@@ -28,11 +27,13 @@ class AnnotationsController < ApplicationController
   # GET /annotations/1.xml
   def show
     @annotation = Annotation.find(params[:id])
-    
     respond_to do |format|
       format.html # show.html.erb
       format.xml  { render :xml => @annotation }
-      format.json { render :json => @annotation }
+      format.json { render :json => @annotation } 
+      format.rdf  {render :rdf => @annotation, :httpURI => base_uri(request.url)}
+      format.ttl  {render :ttl => @annotation, :httpURI => base_uri(request.url)}
+      format.nt  {render :nt => @annotation, :httpURI => base_uri(request.url)}
     end
   end
 
