@@ -5,7 +5,5 @@ class Admin < ActiveRecord::Base
          :recoverable, :rememberable, :trackable, :validatable
 
   # Setup accessible (or protected) attributes for your model
-  attr_accessible :username, :email, :password, :password_confirmation, :remember_me, :about_me
-  
-  validates_uniqueness_of :username
+  attr_accessible :email, :password, :password_confirmation, :remember_me
 end
