@@ -1,3 +1,4 @@
+# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -10,86 +11,69 @@
 #
 # It's strongly recommended to check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20110705110725) do
+ActiveRecord::Schema.define(:version => 20120425084518) do
 
   create_table "admins", :force => true do |t|
-    t.string   "email",                                 :default => "", :null => false
-    t.string   "encrypted_password",     :limit => 128, :default => "", :null => false
+    t.string   "email",                  :default => "", :null => false
+    t.string   "encrypted_password",     :default => "", :null => false
     t.string   "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",                         :default => 0
+    t.integer  "sign_in_count",          :default => 0
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
     t.string   "current_sign_in_ip"
     t.string   "last_sign_in_ip"
     t.string   "username"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                             :null => false
+    t.datetime "updated_at",                             :null => false
   end
 
   add_index "admins", ["email"], :name => "index_admins_on_email", :unique => true
   add_index "admins", ["reset_password_token"], :name => "index_admins_on_reset_password_token", :unique => true
 
   create_table "annotations", :force => true do |t|
-    t.string   "title"
-    t.text     "body"
+    t.string   "body"
     t.integer  "user_id"
     t.integer  "map_id"
     t.string   "wkt_data"
-    t.datetime "created_at"
-    t.datetime "updated_at"
-  end
-
-  create_table "collections", :force => true do |t|
-    t.string   "title"
-    t.text     "description"
-    t.boolean  "public"
-    t.integer  "user_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
-  end
-
-  create_table "collections_maps", :id => false, :force => true do |t|
-    t.integer "collection_id"
-    t.integer "map_id"
+    t.datetime "created_at", :null => false
+    t.datetime "updated_at", :null => false
   end
 
   create_table "control_points", :force => true do |t|
     t.string   "name"
-    t.string   "label"
     t.integer  "user_id"
     t.integer  "map_id"
-    t.string   "geonames_uri"
-    t.decimal  "lat",          :precision => 12, :scale => 10
-    t.decimal  "lng",          :precision => 12, :scale => 10
+    t.string   "geonames_id"
+    t.string   "geonames_label"
+    t.decimal  "lat",            :precision => 12, :scale => 10
+    t.decimal  "lng",            :precision => 12, :scale => 10
     t.decimal  "x"
     t.decimal  "y"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                                     :null => false
+    t.datetime "updated_at",                                     :null => false
   end
 
   create_table "maps", :force => true do |t|
-    t.string   "title",                      :null => false
-    t.text     "description"
-    t.string   "tileset_url"
-    t.integer  "views",       :default => 0
-    t.datetime "map_date"
-    t.boolean  "public"
-    t.integer  "user_id"
+    t.string   "identifier"
+    t.string   "title"
+    t.string   "subject"
     t.integer  "width"
     t.integer  "height"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.string   "author"
+    t.string   "date"
+    t.datetime "created_at", :null => false
+    t.datetime "updated_at", :null => false
   end
 
   create_table "users", :force => true do |t|
-    t.string   "email",                                 :default => "", :null => false
-    t.string   "encrypted_password",     :limit => 128, :default => "", :null => false
+    t.string   "email",                  :default => "", :null => false
+    t.string   "encrypted_password",     :default => "", :null => false
     t.string   "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",                         :default => 0
+    t.integer  "sign_in_count",          :default => 0
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
     t.string   "current_sign_in_ip"
@@ -99,12 +83,10 @@ ActiveRecord::Schema.define(:version => 20110705110725) do
     t.string   "location"
     t.text     "about_me"
     t.datetime "deleted_at"
-    t.integer  "maps_count"
-    t.integer  "collections_count"
     t.integer  "annotations_count"
     t.integer  "control_points_count"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",                             :null => false
+    t.datetime "updated_at",                             :null => false
   end
 
   add_index "users", ["email"], :name => "index_users_on_email", :unique => true
